refactor: command/query 제거
