--- src/services/users/applications/commands/InitializeUser.command.ts
@@ -1,5 +1,0 @@
-export class InitializeUserCommand {
-  constructor(public readonly props: InitializeUserCommandProps) {}
-}
-
-interface InitializeUserCommandProps {}
