--- src/services/users/applications/commands/ReserveTokens.command.ts
@@ -1,26 +1,0 @@
-import { UniqueEntityId } from "~shared/core/domain/UniqueEntityId";
-import { HttpStatusBasedRpcException } from "~shared/filters/exceptions";
-
-import { HttpStatus } from "@nestjs/common";
-
-export class ReserveTokensCommand {
-  constructor(public readonly props: ReserveTokensCommandProps) {
-    this.validate(props);
-  }
-
-  private validate(props: ReserveTokensCommandProps): void {
-    if (props.userId === undefined) {
-      throw new HttpStatusBasedRpcException(HttpStatus.BAD_REQUEST, "userId는 필수입니다.");
-    }
-  }
-}
-
-interface ReserveTokensCommandProps {
-  userId: UniqueEntityId;
-}
-
-export interface ReserveTokensCommandResponse {
-  remainingTokens: number;
-  maxTokens: number;
-  reserved: boolean;
-}
