+++ src/services/users/applications/use-cases/dtos/bind-user-to-auth-user.dto.ts
@@ -2,6 +2,11 @@ import { UseCaseCoreResponse } from "~shared/core/applications/UseCase.response"
 import { AuthUsers } from "~users/domains/auth-users/models/auth-users";
 import { Users } from "~users/domains/users/models/Users";
 
+export interface BindAuthUserToUseUseCaseRequest {
+  user: Users;
+  authUser: AuthUsers;
+}
+
 export interface BindAuthUserToUseUseCaseResponse extends UseCaseCoreResponse {
   user?: Users;
   authUser?: AuthUsers;
