+++ src/services/users/applications/use-cases/dtos/connect-auth-channel.dto.ts
@@ -1,4 +1,6 @@
+import { UseCaseCoreResponse } from "~shared/core/applications/UseCase.response";
 import { UniqueEntityId } from "~shared/core/domain/UniqueEntityId";
+import { AuthUsers } from "~users/domains/auth-users/models/auth-users";
 import { AuthChannel } from "~proto/com/hearlers/v1/model/auth_user_pb";
 
 export interface ConnectAuthChannelUseCaseRequest {
@@ -7,4 +9,8 @@ export interface ConnectAuthChannelUseCaseRequest {
     uniqueId: string;
     authChannel: AuthChannel;
   };
+}
+
+export interface ConnectAuthChannelUseCaseResponse extends UseCaseCoreResponse {
+  authUser?: AuthUsers;
 }
