--- src/services/users/applications/useCases/BindAuthUserToUseUseCase/dto/BindAuthUserToUseUseCase.request.ts
@@ -1,7 +1,0 @@
-import { AuthUsers } from "~users/domains/auth-users/models/auth-users";
-import { Users } from "~users/domains/users/models/Users";
-
-export interface BindAuthUserToUseUseCaseRequest {
-  user: Users;
-  authUser: AuthUsers;
-}
