+++ src/services/users/presentations/message/users-message.controller.ts
@@ -1,22 +1,21 @@
 import { UniqueEntityId } from "~shared/core/domain/UniqueEntityId";
 import { KAFKA_CLIENT } from "~shared/core/infrastructure/Config";
 import { kafkaPayloadToProtoMessage } from "~shared/utils/Proto.utils";
-import { ConsumeTokensCommand } from "~users/applications/commands/ConsumeTokens.command";
-import { UsersCounselMessageCreatedEvent } from "~users/applications/events/CounselMessageCreatedEvents";
+import { UsersCounselMessageCreatedEvent } from "~users/applications/events/counsel-message-created.event";
+import { UsersFacade } from "~users/applications/users.facade";
 import {
   CounselMessageCreatedPayload,
   CounselMessageCreatedPayloadSchema,
 } from "~proto/com/hearlers/v1/message/counsel_pb";
 
 import { Controller, Inject, OnModuleInit } from "@nestjs/common";
-import { CommandBus } from "@nestjs/cqrs";
 import { ClientKafka, EventPattern, Payload } from "@nestjs/microservices";
 
 @Controller()
 export class UsersMessageController implements OnModuleInit {
   constructor(
     @Inject(KAFKA_CLIENT) private readonly kafkaClient: ClientKafka,
-    private readonly commandBus: CommandBus,
+    private readonly usersFacade: UsersFacade,
   ) {}
 
   async onModuleInit() {}
@@ -30,6 +29,6 @@ export class UsersMessageController implements OnModuleInit {
       payload,
       CounselMessageCreatedPayloadSchema,
     );
-    await this.commandBus.execute(new ConsumeTokensCommand({ userId: new UniqueEntityId(convertedPayload.userId) }));
+    await this.usersFacade.consumeTokens(new UniqueEntityId(convertedPayload.userId));
   }
 }
